Fixed compile warnings under Visual Studio 2008 on Windows.
git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@3252 0e98fd9d-079e-457f-9c5c-d482923f975d

+++ src/apps/sfplay/sfplay.cpp
@@ -174,11 +174,11 @@ loadOptions()
 {
   helpopt = cmd_options.getBoolOption("help");
   usageopt = cmd_options.getBoolOption("usage");
-  start = cmd_options.getRealOption("start");
-  length = cmd_options.getRealOption("length");
-  repetitions = cmd_options.getRealOption("repetitions");
+  start = (float)cmd_options.getRealOption("start");
+  length = (float)cmd_options.getRealOption("length");
+  repetitions = (float)cmd_options.getRealOption("repetitions");
   verboseopt = cmd_options.getBoolOption("verbose");
-  gain = cmd_options.getRealOption("gain");
+  gain = (float)cmd_options.getRealOption("gain");
   pluginName = cmd_options.getStringOption("plugin");
   fileName   = cmd_options.getStringOption("filename");
   windowsize = cmd_options.getNaturalOption("windowsize");
